Refactor project to alias original submodule pathing to new location in sys init (very much untested).
