--- ea_airflow_util/dags/callables/variable.py
@@ -1,9 +1,0 @@
-import warnings
-
-warnings.warn(
-    "The 'ea_airflow_util.dags.callables.variable' module is deprecated and will be removed in a future update."
-    "Please update your code to use the new location: 'ea_airflow_util.callables.variable'",
-    DeprecationWarning
-)
-
-from ea_airflow_util.callables.variable import check_variable, update_variable
